auth/requirements.txt
fastapi==0.58.0
uvicorn==0.11.5
passlib==1.7.2
bcrypt==3.1.7
mysqlclient
PyJWT

# for test purposes
flake8==3.8.3
ipdb==0.13.2
ipython==7.15.0
